+++ docs/examples/components/path_raw.py
@@ -3,7 +3,6 @@ from http import HTTPStatus
 from typing import Any, assert_type
 
 import pydantic
-from django.urls import include
 
 from dmr import APIError, Controller
 from dmr.errors import ErrorType
@@ -59,7 +58,7 @@ schema = build_schema(router)
 
 urlpatterns = [
     # Register our router in the final url patterns:
-    path(router.prefix, include((router.urls, 'test_app'), namespace='api')),
+    router.to_urlpatterns(namespace='api'),
     # Add swagger:
     path('docs/openapi.json/', OpenAPIJsonView.as_view(schema), name='openapi'),
 ]