+++ docs/examples/getting_started/urls.py
@@ -1,5 +1,3 @@
-from django.urls import include
-
 # Our `path` is an optimized drop-in replacement of `django.urls.path`:
 from dmr.routing import Router, path
 from examples.getting_started.pydantic_controller import UserController
@@ -18,7 +16,7 @@ router = Router(
 
 # Just a regular `urlpatterns` definition, Django-style:
 urlpatterns = [
-    path(router.prefix, include((router.urls, 'rest_app'), namespace='api')),
+    router.to_urlpatterns(namespace='api'),
 ]
 
 # run: {"controller": "UserController", "method": "post", "body": {"email": "user@wms.org"}, "headers": {"X-API-Consumer": "my-api"}, "url": "/api/user/"}  # noqa: ERA001, E501