+++ CHANGELOG.md
@@ -29,6 +29,8 @@ of requirements for an API to count as public.
 
 - Added support of reusable controllers with `@validate`, #1259
 - Added default value to `prefix` parameter in `Router.__init__`, #1267
+- Added `to_urlpatterns` function to include `Router`
+  instances into `urlpatterns` or other routers, #1262
 
 ### Bugfixes